requirements.txt
Flask>=3.1.0
pdfplumber>=0.11.5
pymupdf>=1.25.5
pytesseract>=0.3.13
Pillow>=11.2.0
